Add padding, margins and align content
This was more challenging. I struggled with
aligning the button to the center. Tried googling
it and using align-content: self; as well as
justify-content. Realised I should not be
trying to change it in the button selector,
moved to the div it was in, then Realised
I should adjust the margin of the button, so did
that. It worked visually. Will see their
solution and adjust if necessary.