Align pink box in middle of blue box
Used justify-items to align the box on the main
axis - horizontal axis in this case - and Used
align-items to center the box on the cross axis -
vertical axis in this case.

Justify items work on the main axis while align
items work on the cross axis. If flex-direction
is in its default state - row -, the cross axis
is vertical and the main axis is horizontal. If
flex-direction is set to column, the main axis
is the vertical axis and the cross axis is the
horizontal axis.

Cross axis is called that because it crosses the
main axis and the main axis is dictated by
flex-direction, because that is the direction in
which the flex items of the flex box will grow and
shrink - horizontal if row, vertical if column.