Adjust spacing and layout for header
Adjusted the elements so that they are evenly spaced
and centered on the cross axis. My result is similar
to how the odin project did it, but not as simple.
It has a more roundabout way, but I will learn to
make it more simplistic as I go along. A key challenge
is understanding how the flex attribute interacts
with different boxes that are nested. But I am
learning, and realising the simple solution is
usually the best solution.