Finished number 4
These flex exercises are kicking my ass. I am
managing to get the same or very close results as
the model solution, but I am constantly in awe
at how simple the solutions are vs. mine. Granted
I am still learning and doing well for a beginner.
I love learning all of this, its teaching me a lot
and we are not even at JS.