Fix corrupted local repo
After completing the 6th flex exercise in foundation
I was unable to access the repo. Googled and turned
out it was corrupted.

I removed the whole repo with 'rm -rf /path/to/repository'
and added the repository. Then I replaced foundations with
the local copy I kept.

Hopefully this solves the problem.